feat(UserApproval): implement endpoint and service method for user approval

+++ fusionX/Services/EventManagerService.cs
@@ -1,6 +1,7 @@
 ﻿using EvenTech.Data;
 using EvenTech.dtos;
 using EvenTech.Models;
+using EvenTech.Models.Constraints;
 using EvenTech.Services.Interfaces;
 using EvenTech.Utils;
 
@@ -37,7 +38,19 @@ namespace EvenTech.Services
         public async Task DeleteAsync(uint id)
         {
             var eventManager = await _context.EventManagers.FindAsync(id);
+            if(eventManager is null) return;
             _context.EventManagers.Remove(eventManager);
+            await _context.SaveChangesAsync();
+        }
+
+        public async Task UserApprovalAsync(uint id, uint attendanceId)
+        {
+            var eventManager = await _context.EventManagers.FindAsync(id)
+                ?? throw new Exception($"Host do not exist! ({id})");
+            var attendance = await _context.Attendances.FindAsync(attendanceId)
+                ?? throw new Exception($"Attendance do not exist! ({id})");
+            attendance.EventManagerId = eventManager.Id;
+            attendance.Status = FeedbackConstraints.Status.Present;
             await _context.SaveChangesAsync();
         }
     }
